Spread optimize
